refactor(access): decompose ContentAccessService into cohesive modules (Codex-2pryk.1.1)
Split the 2399-line ContentAccessService monolith into a thin facade plus
five cohesive internal modules under services/content-access/, preserving
the public ContentAccessService class API and behaviour exactly:

- access-decision.ts   — resolveHasContentAccess + assertStreamingAccess
                         (the accessType decision tree: team/followers/
                         subscribers/paid)
- streaming.ts         — resolveStreamableMedia, buildStreamingResponse,
                         HLS master/variant playlist rewriting
- library.ts           — listUserLibrary
- playback-progress.ts — save/getPlaybackProgress
- r2-signer.ts         — R2Signer interface + DevR2Signer

The facade retains the transaction boundaries and delegates through injected
dependency bundles ({ db, obs, purchaseService, ... }) so each module depends
on a narrow shape rather than the service instance.

Behaviour-preserving: the @codex/access suite is byte-identical and green
(185 passed / 4 skipped, unchanged from origin/dev), @codex/access typecheck
is clean, and there are no new Biome findings (the 3 pre-existing
noNonNullAssertion warnings were moved verbatim into library.ts).
